new changes for loopback
the inst_tb we did with professor did not work as a loop back so I made the new changes to supporut the loop back.
now the setters that send data throught the loop back get the same data how ever right now the fifos do not show the same result!
trying to fix that and after it I will add to this pr and will do a complete iob-soc pr at the end.